+++ xcube/core/normalize.py
@@ -19,6 +19,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+import cftime
 import warnings
 from datetime import datetime
 from typing import Optional, Sequence, Union
@@ -28,12 +29,31 @@ import pandas as pd
 import xarray as xr
 from jdcal import jd2gcal
 
+from xcube.core.timecoord import get_timestamp_from_string
+from xcube.core.timecoord import get_timestamps_from_string
+from xcube.core.verify import assert_cube
 
-def normalize_dataset(ds: xr.Dataset) -> xr.Dataset:
+DatetimeTypes = np.datetime64, cftime.datetime, datetime
+Datetime = Union[np.datetime64, cftime.datetime, datetime]
+
+
+def cubify_dataset(ds: xr.Dataset) -> xr.Dataset:
     """
-    Normalize the geo- and time-coding upon opening the given dataset w.r.t.
-    to a common (CF-compatible) convention used within Cate. This will maximize the compatibility of
-    a dataset for usage with Cate's operations.
+    Normalize the geo- and time-coding upon opening the given dataset w.r.t. a common
+    (CF-compatible) convention.
+
+    Will throw a value error if the dataset could not not be converted to a cube.
+    """
+    ds = normalize_dataset(ds)
+    return assert_cube(ds)
+
+
+def normalize_dataset(ds: xr.Dataset,
+                      reverse_decreasing_lat: bool = False
+                      ) -> xr.Dataset:
+    """
+    Normalize the geo- and time-coding upon opening the given dataset w.r.t. a common
+    (CF-compatible) convention.
 
     That is,
     * variables named "latitude" will be renamed to "lat";
@@ -43,22 +63,85 @@ def normalize_dataset(ds: xr.Dataset) -> xr.Dataset:
     * Remove 2D "lat" and "lon" variables;
     * Two new 1D coordinate variables "lat" and "lon" will be generated from original 2D forms.
 
-    Then, if no coordinate variable time is present but the CF attributes "time_coverage_start" and optionally
-    "time_coverage_end" are given, a scalar time dimension and coordinate variable will be generated.
+    Then, if no coordinate variable time is present but the CF attributes "time_coverage_start"
+    and optionally "time_coverage_end" are given, a scalar time dimension and coordinate variable
+    will be generated.
 
     Finally, it will be ensured that a "time" coordinate variable will be of type *datetime*.
 
     :param ds: The dataset to normalize.
+    :param reverse_decreasing_lat: Whether decreasing latitude values shall be normalized so they
+        are increasing
     :return: The normalized dataset, or the original dataset, if it is already "normal".
     """
+    ds = _normalize_zonal_lat_lon(ds)
     ds = normalize_coord_vars(ds)
     ds = _normalize_lat_lon(ds)
     ds = _normalize_lat_lon_2d(ds)
     ds = _normalize_dim_order(ds)
     ds = _normalize_lon_360(ds)
+    if reverse_decreasing_lat:
+        ds = _reverse_decreasing_lat(ds)
     ds = normalize_missing_time(ds)
     ds = _normalize_jd2datetime(ds)
     return ds
+
+
+def _normalize_zonal_lat_lon(ds: xr.Dataset) -> xr.Dataset:
+    """
+    In case that the dataset only contains lat_centers and is a zonal mean dataset,
+    the longitude dimension created and filled with the variable value of certain latitude.
+    :param ds: some xarray dataset
+    :return: a normalized xarray dataset
+    """
+
+    if 'latitude_centers' not in ds.coords or 'lon' in ds.coords:
+        return ds
+
+    ds_zonal = ds.copy()
+    resolution = (ds.latitude_centers[1].values - ds.latitude_centers[0].values)
+    ds_zonal = ds_zonal.assign_coords(
+        lon=[i + (resolution / 2) for i in np.arange(-180.0, 180.0, resolution)])
+
+    for var in ds_zonal.data_vars:
+        if 'latitude_centers' in ds_zonal[var].dims:
+            ds_zonal[var] = xr.concat([ds_zonal[var] for _ in ds_zonal.lon], 'lon')
+            ds_zonal[var]['lon'] = ds_zonal.lon
+            var_dims = ds_zonal[var].attrs.get('dimensions', [])
+            lat_center_index = var_dims.index('latitude_centers')
+            var_dims.remove('latitude_centers')
+            var_dims.append('lat')
+            var_dims.append('lon')
+            var_chunk_sizes = ds_zonal[var].attrs.get('chunk_sizes', [])
+            lat_chunk_size = var_chunk_sizes[lat_center_index]
+            del var_chunk_sizes[lat_center_index]
+            var_chunk_sizes.append(lat_chunk_size)
+            var_chunk_sizes.append(ds_zonal.lon.size)
+    ds_zonal = ds_zonal.rename_dims({'latitude_centers': 'lat'})
+    ds_zonal = ds_zonal.assign_coords(lat=ds.latitude_centers.values)
+    ds_zonal = ds_zonal.drop('latitude_centers')
+    ds_zonal = ds_zonal.transpose(..., 'lat', 'lon')
+
+    has_lon_bnds = 'lon_bnds' in ds_zonal.coords or 'lon_bnds' in ds_zonal
+    if not has_lon_bnds:
+        lon_values = [[i - (resolution / 2), i + (resolution / 2)] for i in ds_zonal.lon.values]
+        ds_zonal = ds_zonal.assign_coords(lon_bnds=xr.DataArray(lon_values, dims=['lon', 'bnds']))
+    has_lat_bnds = 'lat_bnds' in ds_zonal.coords or 'lat_bnds' in ds_zonal
+    if not has_lat_bnds:
+        lat_values = [[i - (resolution / 2), i + (resolution / 2)] for i in ds_zonal.lat.values]
+        ds_zonal = ds_zonal.assign_coords(lat_bnds=xr.DataArray(lat_values, dims=['lat', 'bnds']))
+
+    ds_zonal.lon.attrs['bounds'] = 'lon_bnds'
+    ds_zonal.lon.attrs['long_name'] = 'longitude'
+    ds_zonal.lon.attrs['standard_name'] = 'longitude'
+    ds_zonal.lon.attrs['units'] = 'degrees_east'
+
+    ds_zonal.lat.attrs['bounds'] = 'lat_bnds'
+    ds_zonal.lat.attrs['long_name'] = 'latitude'
+    ds_zonal.lat.attrs['standard_name'] = 'latitude'
+    ds_zonal.lat.attrs['units'] = 'degrees_north'
+
+    return ds_zonal
 
 
 def _normalize_lat_lon(ds: xr.Dataset) -> xr.Dataset:
@@ -87,7 +170,8 @@ def _normalize_lat_lon_2d(ds: xr.Dataset) -> xr.Dataset:
     """
     Detect 2D 'lat', 'lon' variables that span a equi-rectangular grid. Then:
     Drop original 'lat', 'lon' variables
-    Rename original dimensions names of 'lat', 'lon' variables, usually ('y', 'x'), to ('lat', 'lon').
+    Rename original dimensions names of 'lat', 'lon' variables, usually ('y', 'x'), to
+    ('lat', 'lon').
     Insert new 1D 'lat', 'lon' coordinate variables with dimensions 'lat' and 'lon', respectively.
     :param ds: some xarray dataset
     :return: a normalized xarray dataset, or the original one
@@ -118,8 +202,8 @@ def _normalize_lat_lon_2d(ds: xr.Dataset) -> xr.Dataset:
     equal_lat = np.allclose(lat_data_1, lat_data_2, equal_nan=True)
     equal_lon = np.allclose(lon_data_1, lon_data_2, equal_nan=True)
 
-    # Drop lat lon in any case. If note qual_lat and equal_lon subset_spatial_impl will subsequently
-    # fail with a ValidationError
+    # Drop lat lon in any case. If note qual_lat and equal_lon subset_spatial_impl will
+    # subsequently fail with a ValidationError
 
     ds = ds.drop_vars(['lon', 'lat'])
 
@@ -186,6 +270,24 @@ def _normalize_lon_360(ds: xr.Dataset) -> xr.Dataset:
     return ds.assign(**new_vars)
 
 
+def _reverse_decreasing_lat(ds: xr.Dataset) -> xr.Dataset:
+    """
+    In case the latitude decreases, invert it
+    :param ds: some xarray dataset
+    :return: a normalized xarray dataset
+    """
+    try:
+        if _is_lat_decreasing(ds.lat):
+            ds = ds.sel(lat=slice(None, None, -1))
+    except AttributeError:
+        # The dataset doesn't have 'lat', probably not geospatial
+        pass
+    except ValueError:
+        # The dataset still has an ND 'lat' array
+        pass
+    return ds
+
+
 def _normalize_jd2datetime(ds: xr.Dataset) -> xr.Dataset:
     """
     Convert the time dimension of the given dataset from Julian date to
@@ -240,8 +342,8 @@ def normalize_coord_vars(ds: xr.Dataset) -> xr.Dataset:
     Any data variable is considered a coordinate variable
 
     * whose name is its only dimension name;
-    * whose number of dimensions is two and where the first dimension name is also a variable namd and
-      whose last dimension is named "bnds".
+    * whose number of dimensions is two and where the first dimension name is also a variable name
+        and whose last dimension is named "bnds".
 
     :param ds: The dataset
     :return: The same dataset or a shallow copy with potential coordinate
@@ -262,7 +364,8 @@ def normalize_coord_vars(ds: xr.Dataset) -> xr.Dataset:
 
     old_ds = ds
     ds = old_ds.drop_vars(coord_var_names)
-    ds = ds.assign_coords(**{bounds_var_name: old_ds[bounds_var_name] for bounds_var_name in coord_var_names})
+    ds = ds.assign_coords(**{bounds_var_name: old_ds[bounds_var_name]
+                             for bounds_var_name in coord_var_names})
 
     return ds
 
@@ -270,43 +373,41 @@ def normalize_coord_vars(ds: xr.Dataset) -> xr.Dataset:
 def normalize_missing_time(ds: xr.Dataset) -> xr.Dataset:
     """
     Add a time coordinate variable and their associated bounds coordinate variables
-    if temporal CF attributes ``time_coverage_start`` and ``time_coverage_end``
-    are given but the time dimension is missing.
+    if either temporal CF attributes ``time_coverage_start`` and ``time_coverage_end``
+    are given or time information can be extracted from the file name but the time dimension is
+    missing.
+
+    In case the time information is given by a variable called 't' instead of 'time', it will be
+    renamed into 'time'.
 
     The new time coordinate variable will be named ``time`` with dimension ['time'] and shape [1].
-    The time bounds coordinates variable will be named ``time_bnds`` with dimensions ['time', 'bnds'] and shape [1,2].
+    The time bounds coordinates variable will be named ``time_bnds``
+    with dimensions ['time', 'bnds'] and shape [1,2].
     Both are of data type ``datetime64``.
 
     :param ds: Dataset to adjust
     :return: Adjusted dataset
     """
-    time_coverage_start = ds.attrs.get('time_coverage_start')
-    if time_coverage_start is not None:
-        # noinspection PyBroadException
-        try:
-            time_coverage_start = pd.to_datetime(time_coverage_start)
-        except BaseException:
-            pass
-
-    time_coverage_end = ds.attrs.get('time_coverage_end')
-    if time_coverage_end is not None:
-        # noinspection PyBroadException
-        try:
-            time_coverage_end = pd.to_datetime(time_coverage_end)
-        except BaseException:
-            pass
+    time_coverage_start, time_coverage_end = _get_time_coverage_from_ds(ds)
 
     if not time_coverage_start and not time_coverage_end:
         # Can't do anything
         return ds
 
-    if 'time' in ds:
-        time = ds.time
+    time = _get_valid_time_coord(ds, 'time')
+    if time is None:
+        time = _get_valid_time_coord(ds, 't')
+        if time is not None:
+            ds = ds.rename_vars({"t": "time"})
+            ds = ds.assign_coords(time=('time', ds.time))
+            time = ds.time
+    if time is not None:
         if not time.dims:
             ds = ds.drop_vars('time')
         elif len(time.dims) == 1:
             time_dim_name = time.dims[0]
-            is_time_used_as_dim = any([(time_dim_name in ds[var_name].dims) for var_name in ds.data_vars])
+            is_time_used_as_dim = any([(time_dim_name in ds[var_name].dims)
+                                       for var_name in ds.data_vars])
             if is_time_used_as_dim:
                 # It seems we already have valid time coordinates
                 return ds
@@ -314,7 +415,8 @@ def normalize_missing_time(ds: xr.Dataset) -> xr.Dataset:
             if time_bnds_var_name in ds:
                 ds = ds.drop_vars(time_bnds_var_name)
             ds = ds.drop_vars('time')
-            ds = ds.drop_vars([var_name for var_name in ds.coords if time_dim_name in ds.coords[var_name].dims])
+            ds = ds.drop_vars([var_name for var_name in ds.coords
+                               if time_dim_name in ds.coords[var_name].dims])
 
     try:
         ds = ds.expand_dims('time')
@@ -328,8 +430,9 @@ def normalize_missing_time(ds: xr.Dataset) -> xr.Dataset:
     if time_coverage_start and time_coverage_end:
         has_time_bnds = 'time_bnds' in ds.coords or 'time_bnds' in ds
         if not has_time_bnds:
-            new_coord_vars.update(time_bnds=xr.DataArray([[time_coverage_start, time_coverage_end]],
-                                                         dims=['time', 'bnds']))
+            new_coord_vars.update(
+                time_bnds=xr.DataArray([[time_coverage_start, time_coverage_end]],
+                                       dims=['time', 'bnds']))
     ds = ds.assign_coords(**new_coord_vars)
     ds.coords['time'].attrs['long_name'] = 'time'
     ds.coords['time'].attrs['standard_name'] = 'time'
@@ -341,6 +444,36 @@ def normalize_missing_time(ds: xr.Dataset) -> xr.Dataset:
         ds.coords['time_bnds'].encoding['units'] = 'days since 1970-01-01'
 
     return ds
+
+
+def _get_valid_time_coord(ds: xr.Dataset, name: str) -> Optional[xr.DataArray]:
+    if name in ds:
+        time = ds[name]
+        if time.size > 0 \
+                and time.ndim == 1 \
+                and _is_supported_time_dtype(time.dtype):
+            return time
+    return None
+
+
+def _is_supported_time_dtype(dtype: np.dtype) -> bool:
+    return any(np.issubdtype(dtype, t) for t in DatetimeTypes)
+
+
+def _get_time_coverage_from_ds(ds: xr.Dataset) -> (pd.Timestamp, pd.Timestamp):
+    time_coverage_start = ds.attrs.get('time_coverage_start')
+    if time_coverage_start is not None:
+        time_coverage_start = get_timestamp_from_string(time_coverage_start)
+
+    time_coverage_end = ds.attrs.get('time_coverage_end')
+    if time_coverage_end is not None:
+        time_coverage_end = get_timestamp_from_string(time_coverage_end)
+
+    if time_coverage_start or time_coverage_end:
+        return time_coverage_start, time_coverage_end
+
+    filename = ds.encoding.get('source', '').split('/')[-1]
+    return get_timestamps_from_string(filename)
 
 
 def adjust_spatial_attrs(ds: xr.Dataset, allow_point: bool = False) -> xr.Dataset:
@@ -364,7 +497,7 @@ def adjust_spatial_attrs(ds: xr.Dataset, allow_point: bool = False) -> xr.Datase
     copied = False
 
     for dim in ('lon', 'lat'):
-        geo_spatial_attrs = _get_geo_spatial_cf_attrs_from_var(ds, dim, allow_point=allow_point)
+        geo_spatial_attrs = get_geo_spatial_attrs_from_var(ds, dim, allow_point=allow_point)
         if geo_spatial_attrs:
             # Copy any new attributes into the shallow Dataset copy
             for key in geo_spatial_attrs:
@@ -384,8 +517,9 @@ def adjust_spatial_attrs(ds: xr.Dataset, allow_point: bool = False) -> xr.Datase
         if not copied:
             ds = ds.copy()
 
-        ds.attrs['geospatial_bounds'] = 'POLYGON(({} {}, {} {}, {} {}, {} {}, {} {}))'. \
-            format(lon_min, lat_min, lon_min, lat_max, lon_max, lat_max, lon_max, lat_min, lon_min, lat_min)
+        ds.attrs['geospatial_bounds'] = f'POLYGON(({lon_min} {lat_min}, {lon_min} {lat_max}, ' \
+                                        f'{lon_max} {lat_max}, {lon_max} {lat_min}, ' \
+                                        f'{lon_min} {lat_min}))'
 
         # Determination of the following attributes from introspection in a general
         # way is ambiguous, hence it is safer to drop them than to risk preserving
@@ -397,37 +531,6 @@ def adjust_spatial_attrs(ds: xr.Dataset, allow_point: bool = False) -> xr.Datase
 
         for key in drop:
             ds.attrs.pop(key, None)
-
-    return ds
-
-
-def adjust_temporal_attrs(ds: xr.Dataset) -> xr.Dataset:
-    """
-    Adjust the global temporal attributes of the dataset by doing some
-    introspection of the dataset and adjusting the appropriate attributes
-    accordingly.
-
-    In case the determined attributes do not exist in the dataset, these will
-    be added.
-
-    For more information on suggested global attributes see
-    `Attribute Convention for Data Discovery
-    <http://wiki.esipfed.org/index.php/Attribute_Convention_for_Data_Discovery>`_
-
-    :param ds: Dataset to adjust
-    :return: Adjusted dataset
-    """
-
-    temporal_attrs = _get_temporal_cf_attrs_from_var(ds)
-
-    if temporal_attrs:
-        ds = ds.copy()
-        # Align temporal attributes with the ones from the shallow Dataset copy
-        for key in temporal_attrs:
-            if temporal_attrs[key] is not None:
-                ds.attrs[key] = temporal_attrs[key]
-            else:
-                ds.attrs.pop(key, None)
 
     return ds
 
@@ -447,79 +550,9 @@ def is_bounds_var(ds: xr.Dataset, var: xr.DataArray):
     return False
 
 
-def _get_temporal_cf_attrs_from_var(ds: xr.Dataset, var_name: str = 'time') -> Optional[dict]:
-    """
-    Get temporal boundaries, resolution and duration of the given dataset. If
-    the 'bounds' are explicitly defined, these will be used for calculation,
-    otherwise it will rest on information gathered from the 'time' dimension
-    itself.
-
-    :param ds: Dataset
-    :param var_name: The variable/dimension name.
-    :return: A dictionary {'attr_name': attr_value}
-    """
-
-    if var_name not in ds:
-        return None
-
-    var = ds[var_name]
-
-    if 'bounds' in var.attrs:
-        # According to CF Conventions the corresponding 'bounds' coordinate variable name
-        # should be in the attributes of the coordinate variable
-        bnds_name = var.attrs['bounds']
-    else:
-        # If 'bounds' attribute is missing, the bounds coordinate variable may be named "<dim>_bnds"
-        bnds_name = f'{var_name}_bnds'
-
-    dim_min = dim_max = None
-    dim_var = None
-
-    if bnds_name in ds:
-        bnds_var = ds[bnds_name]
-        if len(bnds_var.shape) == 2 and bnds_var.shape[0] > 0 and bnds_var.shape[1] == 2:
-            dim_var = bnds_var
-            if bnds_var.shape[0] > 1:
-                dim_min = bnds_var.values[0][0]
-                dim_max = bnds_var.values[-1][1]
-            else:
-                dim_min = bnds_var.values[0][0]
-                dim_max = bnds_var.values[0][1]
-
-    if dim_var is None:
-        if len(var.shape) == 1 and var.shape[0] > 0:
-            dim_var = var
-            if var.shape[0] > 1:
-                dim_min = var.values[0]
-                dim_max = var.values[-1]
-            else:
-                dim_min = var.values[0]
-                dim_max = var.values[0]
-
-    # Make sure dim_min and dim_max are valid and are instances of np.datetime64
-    # See https://github.com/CCI-Tools/cate/issues/643
-    if dim_var is None \
-            or not np.issubdtype(dim_var.dtype, np.datetime64):
-        # Cannot determine temporal extent for dimension var_name
-        return None
-
-    if dim_min != dim_max:
-        duration = _get_duration(dim_min, dim_max)
-    else:
-        duration = None
-
-    if dim_min < dim_max and len(var) >= 2:
-        resolution = _get_temporal_res(var.values)
-    else:
-        resolution = None
-
-    return dict(time_coverage_start=str(dim_min),
-                time_coverage_end=str(dim_max),
-                time_coverage_duration=duration,
-                time_coverage_resolution=resolution)
-
-
-def _get_geo_spatial_cf_attrs_from_var(ds: xr.Dataset, var_name: str, allow_point: bool = False) -> Optional[dict]:
+def get_geo_spatial_attrs_from_var(ds: xr.Dataset,
+                                   var_name: str,
+                                   allow_point: bool = False) -> Optional[dict]:
     """
     Get spatial boundaries, resolution and units of the given dimension of the given
     dataset. If the 'bounds' are explicitly defined, these will be used for
@@ -536,13 +569,18 @@ def _get_geo_spatial_cf_attrs_from_var(ds: xr.Dataset, var_name: str, allow_poin
         return None
 
     var = ds[var_name]
+    res_name = 'geospatial_{}_resolution'.format(var_name)
+    min_name = 'geospatial_{}_min'.format(var_name)
+    max_name = 'geospatial_{}_max'.format(var_name)
+    units_name = 'geospatial_{}_units'.format(var_name)
 
     if 'bounds' in var.attrs:
         # According to CF Conventions the corresponding 'bounds' coordinate variable name
         # should be in the attributes of the coordinate variable
         bnds_name = var.attrs['bounds']
     else:
-        # If 'bounds' attribute is missing, the bounds coordinate variable may be named "<dim>_bnds"
+        # If 'bounds' attribute is missing, the bounds coordinate variable may be named
+        # "<dim>_bnds"
         bnds_name = '%s_bnds' % var_name
 
     dim_var = None
@@ -551,27 +589,47 @@ def _get_geo_spatial_cf_attrs_from_var(ds: xr.Dataset, var_name: str, allow_poin
         bnds_var = ds[bnds_name]
         if len(bnds_var.shape) == 2 and bnds_var.shape[0] > 0 and bnds_var.shape[1] == 2:
             dim_var = bnds_var
-            dim_res = abs(bnds_var.values[0][1] - bnds_var.values[0][0])
+            dim_res = abs(bnds_var[0, 1] - bnds_var[0, 0])
             if bnds_var.shape[0] > 1:
-                dim_min = min(bnds_var.values[0][0], bnds_var.values[-1][1])
-                dim_max = max(bnds_var.values[0][0], bnds_var.values[-1][1])
+                dim_min = min(bnds_var[0, 0], bnds_var[-1, 1])
+                dim_max = max(bnds_var[0, 0], bnds_var[-1, 1])
             else:
-                dim_min = min(bnds_var.values[0][0], bnds_var.values[0][1])
-                dim_max = max(bnds_var.values[0][0], bnds_var.values[0][1])
+                dim_min = min(bnds_var[0, 0], bnds_var[0, 1])
+                dim_max = max(bnds_var[0, 0], bnds_var[0, 1])
 
     if dim_var is None:
         if len(var.shape) == 1 and var.shape[0] > 0:
             if var.shape[0] > 1:
                 dim_var = var
-                dim_res = abs(var.values[1] - var.values[0])
-                dim_min = min(var.values[0], var.values[-1]) - 0.5 * dim_res
-                dim_max = max(var.values[0], var.values[-1]) + 0.5 * dim_res
-            elif len(var.values) == 1 and allow_point:
-                dim_var = var
-                # Actually a point with no extent
-                dim_res = 0.0
-                dim_min = var.values[0]
-                dim_max = var.values[0]
+                dim_res = abs(var[1] - var[0])
+                dim_min = min(var[0], var[-1]) - 0.5 * dim_res
+                dim_max = max(var[0], var[-1]) + 0.5 * dim_res
+            elif var.size == 1:
+                if res_name in ds.attrs:
+                    dim_res = ds.attrs[res_name]
+                    if isinstance(dim_res, str):
+                        # remove any units from string
+                        dim_res = dim_res.replace('degree', '').replace('deg', '').\
+                            replace('°', '').strip()
+                    try:
+                        dim_res = float(dim_res)
+                        dim_var = var
+                        # Consider extent in metadata if provided
+                        dim_min = var[0] - 0.5 * dim_res
+                        dim_max = var[0] + 0.5 * dim_res
+                    except ValueError:
+                        if allow_point:
+                            dim_var = var
+                            # Actually a point with no extent
+                            dim_res = 0.0
+                            dim_min = var[0]
+                            dim_max = var[0]
+                elif allow_point:
+                    dim_var = var
+                    # Actually a point with no extent
+                    dim_res = 0.0
+                    dim_min = var[0]
+                    dim_max = var[0]
 
     if dim_var is None:
         # Cannot determine spatial extent for variable/dimension var_name
@@ -581,11 +639,6 @@ def _get_geo_spatial_cf_attrs_from_var(ds: xr.Dataset, var_name: str, allow_poin
         dim_units = var.attrs['units']
     else:
         dim_units = None
-
-    res_name = 'geospatial_{}_resolution'.format(var_name)
-    min_name = 'geospatial_{}_min'.format(var_name)
-    max_name = 'geospatial_{}_max'.format(var_name)
-    units_name = 'geospatial_{}_units'.format(var_name)
 
     geo_spatial_attrs = dict()
     # noinspection PyUnboundLocalVariable
@@ -597,40 +650,6 @@ def _get_geo_spatial_cf_attrs_from_var(ds: xr.Dataset, var_name: str, allow_poin
     geo_spatial_attrs[units_name] = dim_units
 
     return geo_spatial_attrs
-
-
-def _get_temporal_res(time: np.ndarray) -> str:
-    """
-    Determine temporal resolution of the given datetimes array.
-
-    See also: `ISO 8601 Durations <https://en.wikipedia.org/wiki/ISO_8601#Durations>`_
-
-    :param time: A numpy array containing np.datetime64 objects
-    :return: Temporal resolution formatted as an ISO 8601:2004 duration string
-    """
-    delta = time[1] - time[0]
-    days = delta.astype('timedelta64[D]') / np.timedelta64(1, 'D')
-
-    if (27 < days) and (days < 32):
-        return 'P1M'
-    else:
-        return 'P{}D'.format(int(days))
-
-
-def _get_duration(tmin: np.datetime64, tmax: np.datetime64) -> str:
-    """
-    Determine the duration of the given datetimes.
-
-    See also: `ISO 8601 Durations <https://en.wikipedia.org/wiki/ISO_8601#Durations>`_
-
-    :param tmin: Time minimum
-    :param tmax: Time maximum
-    :return: Temporal resolution formatted as an ISO 8601:2004 duration string
-    """
-    delta = tmax - tmin
-    day = np.timedelta64(1, 'D')
-    days = (delta.astype('timedelta64[D]') / day) + 1
-    return 'P{}D'.format(int(days))
 
 
 def get_lon_dim_name_impl(ds: Union[xr.Dataset, xr.DataArray]) -> Optional[str]:
@@ -651,7 +670,8 @@ def get_lat_dim_name_impl(ds: Union[xr.Dataset, xr.DataArray]) -> Optional[str]:
     return _get_dim_name(ds, ['lat', 'latitude'])
 
 
-def _get_dim_name(ds: Union[xr.Dataset, xr.DataArray], possible_names: Sequence[str]) -> Optional[str]:
+def _get_dim_name(ds: Union[xr.Dataset, xr.DataArray], possible_names: Sequence[str]) \
+        -> Optional[str]:
     for name in possible_names:
         if name in ds.dims:
             return name
@@ -662,7 +682,7 @@ def _is_lat_decreasing(lat: xr.DataArray) -> bool:
     """
     Determine if the latitude is decreasing
     """
-    if lat.values[0] > lat.values[-1]:
+    if lat[0] > lat[-1]:
         return True
 
     return False
@@ -698,7 +718,7 @@ def _normalize_dim_order(ds: xr.Dataset) -> xr.Dataset:
                 ds = ds.copy()
                 copy_created = True
             ds[var_name] = var.transpose(*dim_names)
-            if var.encoding:
+            if var.encoding and 'chunksize' in ds[var_name].data:
                 ds[var_name].encoding['chunks'] = ds[var_name].data.chunksize
 
     return ds